tbf_dl: uint8_t is enough to store a TA value
According to 3GPP TS 44.018 sec 10.5.2.40, Timing Advance value is 8 bit
and range is 0-63 (0-219 on GSM400). So there's no need for 16 bits to
store it. uint8_t is used in all other places in the code.

+++ src/tbf_dl.cpp
@@ -138,7 +138,7 @@ static int tbf_new_dl_assignment(struct gprs_rlcmac_bts *bts,
 {
 	bool ss;
 	int8_t use_trx;
-	uint16_t ta = GSM48_TA_INVALID;
+	uint8_t ta = GSM48_TA_INVALID;
 	struct gprs_rlcmac_ul_tbf *ul_tbf = NULL, *old_ul_tbf;
 	struct gprs_rlcmac_dl_tbf *dl_tbf = NULL;
 	GprsMs *ms;
